feat: add `?tp_ignore_side_effects_in_prod/2`
Sometimes, we might use expensive calls inside `?tp/2` for testing
purposes.  But, the tracepoint is in a hot path, then the expensive
call will still be called.

For those cases, we introduce the `?tp_ignore_side_effects_in_prod/2`
macro to give users the possibility of optimizing their specific use
cases if necessary.